php code for viewing pdf database files
The goal is to build an application to view pdfs in a viewbox with the ability to scroll through the pdf, rotate pages if they are upside down, zoom in or out, and maybe darken the background.  I do not, however want the viewer to be able to print on this one.  I wouldn't mind having a version where the viewer can print and a version where they cannot.